Fixes -Wall warnings and enables -Wall

+++ vm/core.cpp
@@ -116,7 +116,7 @@ Value read_file(Value fileName)
     char* buf = (char*)malloc(len+1);
 
     // Read into the allocated buffer
-    int read = fread(buf, 1, len, file);
+    size_t read = fread(buf, 1, len, file);
 
     if (read != len)
     {
